move neo4j.service.ts
