+++ src/neo4j/neo4j.module.ts
@@ -1,5 +1,5 @@
 import { DynamicModule, Module, Provider } from "@nestjs/common";
-import { Neo4jService } from "./neo4j.service";
+import { Neo4jService } from "./services/neo4j.service";
 import { NEO4J_DRIVER, NEO4J_OPTIONS } from "./neo4j.constants";
 import { createDriver } from "./neo4j.utils";
 import { Neo4jConfig } from "./neo4jConfig.interface";
