+++ src/users/services/users.service.ts
@@ -1,7 +1,7 @@
 import { Inject, Injectable } from "@nestjs/common";
 import { RegisterUserPayloadDto, Role, User } from "../models";
 import { IUsersService } from "./users.service.interface";
-import { Neo4jService } from "../../neo4j/neo4j.service";
+import { Neo4jService } from "../../neo4j/services/neo4j.service";
 import { v4 as uuidv4 } from "uuid";
 
 @Injectable()
